requirements.txt
fastapi~=0.128.0
python-dotenv~=1.2.2
RapidFuzz~=3.14.3
requests~=2.32.5
urllib3~=2.6.3
pypdf~=6.9.2
rich~=14.3.2
aiohttp~=3.13.5
json_repair~=0.58.7
EbookLib~=0.20